publicly accessible Profile Picture
https://azurefreestorage1.blob.core.windows.net/profilepictures/9a5851d7-b96a-471b-9da9-f70cb8db3856koushik.png

<Error> <Code>PublicAccessNotPermitted</Code> <Message>Public access is not permitted on this storage account. RequestId:b0fd82ca-401e-003c-1ba5-bcb93c000000 Time:2024-06-12T08:52:16.4017025Z</Message> </Error>
===============================
o solve the error "The access level is set to private because anonymous access is disabled on this storage account," you need to modify the access level of the storage account to allow public read access.

Here are the steps to do so:

Using Azure CLI

You can use the Azure CLI to create a storage account with the allowBlobPublicAccess property set to true. Here's an example:
=======================
az storage account create \
    --name <storage-account> \
    --resource-group <resource-group> \
    --kind StorageV2 \
    --location <location> \
    --allow-blob-public-access true
===========================
Using Azure Portal

You can also modify the access level of the storage account using the Azure Portal. Here's how:

Go to the Azure Portal and navigate to your storage account.
Click on the "Containers" blade under the "Data Storage" section.
Select the container that contains the blob (in this case, "profilepictures").
Click on the "Change access level" button under the "Settings" section.
Select "Container (anonymous read access for containers and blobs)" and click "Save".
After you've made these changes, the profile picture should be accessible to the public using the URL you provided.

Note that making the container publicly accessible can have security implications, so make sure to only grant public access to the minimum set of resources necessary.

Edit
===================================
The access level is set to private because anonymous access is disabled on this storage account
===================
https://www.youtube.com/watch?v=mJA7aAtXjws